+++ src/components/@common/Fallback/Empty/EmptyFallback.tsx
@@ -1,6 +1,6 @@
 import Description from "../../Description/Description";
-import EmptyCart from "/planet-empty-cart.svg";
 import * as S from "./EmptyFallback.styles";
+import EmptyCart from "/planet-empty-cart.svg";
 
 const EmptyFallback = () => {
   return (
